LayoutTests/inspector/console/console-url-line-column-expected.txt

CONSOLE MESSAGE: line 9: one
CONSOLE MESSAGE: line 10: two
CONSOLE MESSAGE: line 11: three
CONSOLE MESSAGE: line 14: error
CONSOLE MESSAGE: line 14: log
CONSOLE MESSAGE: line 15: info
CONSOLE MESSAGE: line 15: warn
CONSOLE MESSAGE: line 16: assert
Tests that url, line, and column properties are defined for console functions.

Message at location: console-url-line-column.html:9:15
Message at location: console-url-line-column.html:10:15
Message at location: console-url-line-column.html:11:15
Message at location: console-url-line-column.html:14:17
Message at location: console-url-line-column.html:14:39
Message at location: console-url-line-column.html:15:16
Message at location: console-url-line-column.html:15:38
Message at location: console-url-line-column.html:16:18
